% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLPplus_daVinci.R
\name{Chr1Centromere_Heatmap}
\alias{Chr1Centromere_Heatmap}
\title{Title Chr1Centromere_Heatmap}
\usage{
Chr1Centromere_Heatmap(
  HMMcopy_StateIdeal_RmvArtifactCNV_ByMeanGini_File,
  HMMcopy_StateIdeal_RmvArtifactCNV_ByMeanGini_StateLowHighReorderFile,
  MySampleID
)
}
\arguments{
\item{HMMcopy_StateIdeal_RmvArtifactCNV_ByMeanGini_File}{HMMcopy data after remove artifact cells}

\item{HMMcopy_StateIdeal_RmvArtifactCNV_ByMeanGini_StateLowHighReorderFile}{HMMcopy data after remove artifact cells and after reordered}

\item{MySampleID}{Manually typed in.}
}
\value{
CompHeatmap_Chr1CentromereRegion_ pdf file
}
\description{
Title Chr1Centromere_Heatmap
}
\examples{
Chr1Centromere_Heatmap(HMMcopy_StateIdeal_RmvArtifactCNV_ByMeanGini_File="HMMcopy_StateIdeal_RmvArtifactCNV_ByMeanGini_01_213_143839A.rds", HMMcopy_StateIdeal_RmvArtifactCNV_ByMeanGini_StateLowHighReorderFile="HMMcopy_StateIdeal_AfterRmvArtifactCell_RmvReorderCluster_01_213_143839A.rds", MySampleID="01_213_143839A")
}
