% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLPplus_daVinci.R
\name{CNVStatePloidyLine}
\alias{CNVStatePloidyLine}
\title{Title}
\usage{
CNVStatePloidyLine(
  CNVSeg_StateIdealSelect_RmvByGiniMeanState_RmvByCluster,
  TotalSumThreshold = 30
)
}
\arguments{
\item{CNVSeg_StateIdealSelect_RmvByGiniMeanState_RmvByCluster}{CNA state data after removing outliers by High Gini, high mean and cluster number}

\item{TotalSumThreshold}{30}
}
\description{
Title
}
\examples{
 CNVStatePloidyLine(CNVSeg_StateIdealSelect_RmvByGiniMeanState_RmvByCluster, TotalSumThreshold=30)
}
