% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLPplus_daVinci.R
\name{CNVStateProcess}
\alias{CNVStateProcess}
\title{Title CNVStateProcess}
\usage{
CNVStateProcess(
  HMMcopyReadFile,
  RemoveArtifactCell = TRUE,
  BinRegionDelete = TRUE,
  MySampleID = "MySampleID"
)
}
\arguments{
\item{HMMcopyReadFile}{original HMMcopy output file}

\item{RemoveArtifactCell}{boolean}

\item{BinRegionDelete}{boolean}

\item{MySampleID}{manually typed in.}
}
\value{
Processed CNV state data for making hierarchical clustering
}
\description{
Title CNVStateProcess
}
\examples{
CNVStateProcess(HMMcopyReadFile, RemoveArtifactCell, BinRegionDelete, MySampleID)
}
