% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLPplus_daVinci.R
\name{CountCell_EachCluster}
\alias{CountCell_EachCluster}
\title{Title}
\usage{
CountCell_EachCluster(HierarchialClusterInfoFile, ClusterReorder)
}
\arguments{
\item{HierarchialClusterInfoFile}{cluster number information of each cell}

\item{ClusterReorder}{cluster reorder numbers}
}
\value{
IndexCell_ReorderClusterAll
}
\description{
Title
}
\examples{
CountCell_EachCluster(HierarchialClusterInfoFile="RowSplitHeatmapDraw_WithoutArtifactLowHighState7_AllChr_01_213_143839A.rds", ClusterReorder=c(5,3,4,7,6))
}
