% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLPplus_daVinci.R
\name{HierarchialClusteringHeatmap}
\alias{HierarchialClusteringHeatmap}
\title{Title}
\usage{
HierarchialClusteringHeatmap(
  Processed_HMMcopyStateData = HMMcopy_StateIdeal_RmvArtifactCNV_WithoutBinRegion,
  MySampleID = "MySampleID",
  AllCell = "WithArtifact",
  ChromosomeToSubset = "All",
  ClusterSize = 9
)
}
\arguments{
\item{Processed_HMMcopyStateData}{HMMcopyStateData processed from Step1}

\item{MySampleID}{Manually typed in}

\item{AllCell}{With or without artifact cells}

\item{ChromosomeToSubset}{all chromosome}

\item{ClusterSize}{arbitrary cluster size.}
}
\value{
save RowSplitHeatmapDraw_WithArtifact rds file
}
\description{
Title
}
\examples{
HierarchialClusteringHeatmap(Processed_HMMcopyStateData=HMMcopy_AfterRmvCell_WithoutNABinRegion, MySampleID="01_206_143839A", AllCell="WithoutArtifact", ChromosomeToSubset="All",ClusterSize=9)
}
