% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLPplus_daVinci.R
\name{Karyoplote_BySegment}
\alias{Karyoplote_BySegment}
\title{Title}
\usage{
Karyoplote_BySegment(
  CNVSegmentFile =
    "/Users/lees130/Library/CloudStorage/OneDrive-NYULangoneHealth/N07_DLPplus_scWGS/04_DLPplus_1qGain_8ptData/01-213_143839A/hmmcopy_segments.csv.gz",
  HierarchialClusterInfoFile =
    "/Users/lees130/Library/CloudStorage/OneDrive-NYULangoneHealth/N07_DLPplus_scWGS/09b_HMMCopyHierarchicalCluster_Rpackage/02_OutHMMcopyClusteringHeatmap_RmvBinCellByGiniMean/RowSplitHeatmapDraw_WithoutArtifact_AllChr_01_213_143839A.rds",
  HierarchialClusterInfoAfRmvClusterFile =
    "/Users/lees130/Library/CloudStorage/OneDrive-NYULangoneHealth/N07_DLPplus_scWGS/09b_HMMCopyHierarchicalCluster_Rpackage/02_OutHMMcopyClusteringHeatmap_RmvBinCellByGiniMean/RowSplitHeatmapDraw_WithoutArtifactByCluster_AllChr_01_213_143839A.rds",
  HierarchialClusterInfoAfRmvClusterChr1File = NULL,
  CNVStateData_AfterRmvHighGiniMeanFile =
    "/Users/lees130/Library/CloudStorage/OneDrive-NYULangoneHealth/N07_DLPplus_scWGS/09b_HMMCopyHierarchicalCluster_Rpackage/02_OutHMMcopyClusteringHeatmap_RmvBinCellByGiniMean/HMMcopy_StateIdeal_RmvCellByGiniMean_WithoutBinRegion_01_213_143839A.rds",
  CNVStateData_AfterRmvHighGiniMeanClusterFile = NULL,
  ClusterNumbToRmv = c(7, 8, 9),
  ClusterNumbToPick = c(6, 8),
  ClusterNumbToPick_InChr1 = c(1, 2),
  MySampleID = "01_213_143839A"
)
}
\arguments{
\item{CNVSegmentFile}{original CNV segment file}

\item{HierarchialClusterInfoFile}{cluster number of cells}

\item{HierarchialClusterInfoAfRmvClusterFile}{cluster number of cells after removing outliers}

\item{HierarchialClusterInfoAfRmvClusterChr1File}{cluster number of cells in chr 1}

\item{CNVStateData_AfterRmvHighGiniMeanFile}{CNV state data after removing cells of high gini and high mean}

\item{CNVStateData_AfterRmvHighGiniMeanClusterFile}{CNV state data after removing cells of high gini and high mean, and specific clusters}

\item{ClusterNumbToRmv}{cluster number to remove, NULL}

\item{ClusterNumbToPick}{cluster number to make ploidy plot}

\item{ClusterNumbToPick_InChr1}{cluster number to make ploidy plots in chr1}

\item{MySampleID}{manually typed in}
}
\description{
Title
}
\examples{
  Karyoplote_BySegment(CNVSegmentFile,HierarchialClusterInfoFile,HierarchialClusterInfoAfRmvClusterFile, HierarchialClusterInfoAfRmvClusterChr1File,CNVStateData_AfterRmvHighGiniMeanFile,CNVStateData_AfterRmvHighGiniMeanClusterFile, ClusterNumbToRmv,  ClusterNumbToPick,ClusterNumbToPick_InChr1, MySampleID)
}
