% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLPplus_daVinci.R
\name{PhylogeneticTree}
\alias{PhylogeneticTree}
\title{Title PhylogeneticTree}
\usage{
PhylogeneticTree(
  df = TreeData,
  num_cells_in = NumbCellData,
  MySampleID = MySampleID
)
}
\arguments{
\item{df}{data.frame of tree structure}

\item{num_cells_in}{data.frame of cell numbers}

\item{MySampleID}{my sample ID}
}
\description{
Title PhylogeneticTree
}
\examples{
PhylogeneticTree(df=TreeData, num_cells_in=NumbCellData, MySampleID=MySampleID)
}
