% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLPplus_daVinci.R
\name{plotLRR}
\alias{plotLRR}
\title{Title}
\usage{
plotLRR(
  karyoplot,
  snps,
  lrr.column = "lrr",
  labels = "LRR",
  ymin = -4,
  ymax = 2,
  out.of.range = "points",
  out.of.range.col = "white",
  density.height = 0.05,
  density.window = 1e+05,
  line.at.0 = TRUE,
  line.at.0.col = "blue",
  r0 = 0,
  r1 = 1,
  points.cex = 0.3,
  points.col = "#333333",
  points.pch = 16,
  label.cex = 1.5,
  label.srt = 90,
  label.margin = 0.03,
  add.axis = TRUE,
  axis.cex = 1.2,
  track.margin = 0.1,
  data.panel = 1,
  verbose = FALSE
)
}
\arguments{
\item{karyoplot}{karyoplot obejct}

\item{snps}{karyoplot tiles}

\item{lrr.column}{lrr}

\item{labels}{LRR}

\item{ymin}{-4}

\item{ymax}{2}

\item{out.of.range}{points}

\item{out.of.range.col}{white}

\item{density.height}{0.05}

\item{density.window}{1e+05}

\item{line.at.0}{TRUE}

\item{line.at.0.col}{blue}

\item{r0}{0}

\item{r1}{1}

\item{points.cex}{0.3}

\item{points.col}{#333333}

\item{points.pch}{16}

\item{label.cex}{1.5}

\item{label.srt}{90}

\item{label.margin}{0.03}

\item{add.axis}{TRUE}

\item{axis.cex}{1.2}

\item{track.margin}{0.1}

\item{data.panel}{1}

\item{verbose}{Boolean}
}
\description{
Title
}
\examples{
  plotLRR(karyoplot, snps, lrr.column, labels, ymin, ymax, out.of.range, out.of.range.col,density.height, density.window, line.at.0, line.at.0.col, r0, r1, points.cex, points.col, points.pch, label.cex, label.srt, label.margin, add.axis, axis.cex,track.margin, data.panel, verbose)
}
