% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLPplus_daVinci.R
\name{RemoveArtifactCellHeatmap}
\alias{RemoveArtifactCellHeatmap}
\title{Title RemoveArtifactCellHeatmap}
\usage{
RemoveArtifactCellHeatmap(
  CNVStateData_AfterRmvHighGiniMeanFile =
    "HMMcopy_StateIdeal_RmvCellByGiniMean_WithoutBinRegion_01_213_143939A.rds",
  MySampleID,
  KM_ClusterSize = 9,
  ThresholdLowState = 0.2,
  ThresholdHighState = 10
)
}
\arguments{
\item{CNVStateData_AfterRmvHighGiniMeanFile}{Output of step2.}

\item{MySampleID}{Manually typed in}

\item{KM_ClusterSize}{cluster size that I want}

\item{ThresholdLowState}{bins of low state are bigger than 20%}

\item{ThresholdHighState}{bins of high state are more than 10}
}
\value{
make CompHmapCNV_RmvByGiniMeanLowHighState_AllChr pdf file
}
\description{
Title RemoveArtifactCellHeatmap
}
\examples{
RemoveArtifactCellHeatmap(CNVStateData_AfterRmvHighGiniMeanFile="HMMcopy_StateIdeal_RmvCellByGiniMean_WithoutBinRegion_01_213_143939A.rds",MySampleID, KM_ClusterSize=9, Chr1_KM_ClusterSize=5, ThresholdLowState)
}
