% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLPplus_daVinci.R
\name{RemoveNormalClusterReorderHeatmap}
\alias{RemoveNormalClusterReorderHeatmap}
\title{Title}
\usage{
RemoveNormalClusterReorderHeatmap(
  HierarchialClusterInfoFile =
    "RowSplitHeatmapDraw_WithoutArtifactLowHighState6_AllChr_01_213_143839A.rds",
  CNVStateData_AfterRmvHighGiniMeanFile =
    "HMMcopy_StateIdeal_RmvCellByGiniMean_WithoutBinRegion_01_213_143939A.rds",
  MySampleID,
  ClusterNumbToRmv = c(1, 2),
  ClusterReorder = c(3, 6, 4, 5)
)
}
\arguments{
\item{HierarchialClusterInfoFile}{cluster numbers of each cell}

\item{CNVStateData_AfterRmvHighGiniMeanFile}{CNV data after removing cells of high Gini score and high mean.}

\item{MySampleID}{Manually typed in}

\item{ClusterNumbToRmv}{cluster number to remove}

\item{ClusterReorder}{cluster number to reorder}
}
\value{
make CompHmapCNV_RmvByGiniMeanLowHighState_LowSt_Chr1only_ pdf file
}
\description{
Title
}
\examples{
RemoveNormalClusterReorderHeatmap(HierarchialClusterInfoFile="RowSplitHeatmapDraw_WithoutArtifactLowHighState6_AllChr_01_213_143839A.rds",CNVStateData_AfterRmvHighGiniMeanFile="HMMcopy_StateIdeal_RmvCellByGiniMean_WithoutBinRegion_01_213_143939A.rds", MySampleID="01_213_143839A",ClusterNumbToRmv=c(1,2), ClusterReorder=c(3,6,4,5)
}
