% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLPplus_daVinci.R
\name{RemoveReorderCell_ByClusterNumb}
\alias{RemoveReorderCell_ByClusterNumb}
\title{Title}
\usage{
RemoveReorderCell_ByClusterNumb(
  HMMcopy_StateIdeal_AfterRmvArtifactCell_ColNameProc,
  HierarchialClusterInfoFile,
  ClusterNumbToRmv = NULL,
  ClusterReorder = NULL,
  ClusterNumbToPick = NULL,
  MySampleID,
  ForPloidyPlot = FALSE
)
}
\arguments{
\item{HMMcopy_StateIdeal_AfterRmvArtifactCell_ColNameProc}{Output file of Step3. Input for Step4.}

\item{HierarchialClusterInfoFile}{cluster numbers for each cell}

\item{ClusterNumbToRmv}{cluster number that I want to remove}

\item{ClusterReorder}{Cluster reorder number}

\item{ClusterNumbToPick}{If I want to pick a couple of clusters.}

\item{ForPloidyPlot}{Is RemoveReorderCell_ByClusterNumb is for making heatmap or ploidyplot?}
}
\value{
HMMcopy_StateIdeal_AfterRmvArtifactCell_ColNameProc
}
\description{
Title
}
\examples{
RemoveReorderCell_ByClusterNumb(HMMcopy_StateIdeal_AfterRmvArtifactCell_ColNameProc, HierarchialClusterInfoFile, ClusterNumbToRmv=NULL,ClusterReorder=NULL, ClusterNumbToPick=NULL)
}
